requirements.txt
# Drip SDK Health Check CLI - Python version dependencies

# Drip SDK from PyPI
drip-sdk>=0.1.0

# CLI framework
click>=8.0.0

# Environment variable management
python-dotenv>=1.0.0
